Finised battles and created battle tests
battle is done and now lets you battle, reset both armys health and battle again with the same armys, or generate completly new armys and battle again.

Battle test tests a large number or battles that are battle after the same aryms are reset or after new armys are generated.
From these tests I found that sometimes a army is generated much stornger then the other and wins almost all the time if you just keep reseting them and battleing again or if the armys are very close when you reset and battle a bunch of times you will get close to the same ammount of wins and losses and some draws.

I also fixed the creature class so that the hp and full health are the same before they would have been two diffrent numbers.